Fix memory leak
Using the simulink gym interface caused high memory consumption. It was
found that this is caused by using Matlab functionality for setting
block, workspace and model parameters of the SimulationInput object in
every reset of the environment. This commit gets rid of the behavior in
the reset method. It has to be investigated if the memory consumption of
these methods improves with newer Matlab version (>R2020b).